// src/types/customer.ts
export interface ICustomer {
  CustomerID: number;
  Code: string[];
  Name: string[];
  Contact: string[];
  Telephone: string[];
  Mobile: string[];
  Fax: string[];
  Email: string[];
  Address1: string[];
  Address2: string[];
  Address3: string[];
  Address4: string[];
  CountryName: string[];
  CountryCode: string[];
  Postcode: string[];
  Website: string[];
  EC: number;
  OutsideEC: number;
  Notes: string[];
  Source: number;
  Discount: string;
  ShowDiscount: string;
  PaymentTerms: number;
  ExtraText1: string[];
  ExtraText2: string[];
  ExtraText3: string[];
  ExtraText4: string[];
  ExtraText5: string[];
  ExtraText6: string[];
  ExtraText7: string[];
  ExtraText8: string[];
  ExtraText9: string[];
  ExtraText10: string[];
  ExtraText11: string[];
  ExtraText12: string[];
  ExtraText13: string[];
  ExtraText14: string[];
  ExtraText15: string[];
  ExtraText16: string[];
  ExtraText17: string[];
  ExtraText18: string[];
  ExtraText19: string[];
  ExtraText20: string[];
  CheckBox1: number;
  CheckBox2: number;
  CheckBox3: number;
  CheckBox4: number;
  CheckBox5: number;
  CheckBox6: number;
  CheckBox7: number;
  CheckBox8: number;
  CheckBox9: number;
  CheckBox10: number;
  CheckBox11: number;
  CheckBox12: number;
  CheckBox13: number;
  CheckBox14: number;
  CheckBox15: number;
  CheckBox16: number;
  CheckBox17: number;
  CheckBox18: number;
  CheckBox19: number;
  CheckBox20: number;
  Created: string;
  Updated: string;
  CurrencyID: number;
  ContactTitle: string[];
  ContactFirstName: string[];
  ContactLastName: string[];
  CustHasDeliveryAddress: string[];
  DeliveryAddress1: string[];
  DeliveryAddress2: string[];
  DeliveryAddress3: string[];
  DeliveryAddress4: string[];
  DeliveryCountryName: string[];
  DeliveryCountryCode: string[];
  DeliveryPostcode: string[];
  VATNumber: string[];
}